% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{py_builtins}
\alias{py_builtins}
\title{Python's built-in functions}
\usage{
py_builtins
}
\description{
These built-in functions are imported when \code{py_builtins} is first evaluated
and Python is already initialized.
}
