% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{reimport_py_builtins}
\alias{reimport_py_builtins}
\title{Re-import Python's built-in functions}
\usage{
reimport_py_builtins(convert = FALSE)
}
\arguments{
\item{convert}{Boolean. Whether to automatically convert Python objects to R. See also \code{\link[reticulate:import]{reticulate::import()}}.}
}
\description{
This function re-imports Python's built-in functions into \link{py_builtins}
by calling \code{\link[reticulate:import]{reticulate::import_builtins()}}.
}
\examples{

\dontrun{
reimport_py_builtins()
}

}
