% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{py_builtins}
\alias{py_builtins}
\title{Python's built-in functions}
\usage{
py_builtins
}
\description{
These built-in functions are imported when the package loads.
By default, \code{reticulate::py_discover_config()} is used to determine the Python version.
To use a different Python interpreter, first set it with \code{reticulate::use_python()},
then re-import the built-in functions by calling \code{reimport_py_builtins()}.
}
