% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_class.R
\name{py_class}
\alias{py_class}
\title{Create a Python class}
\usage{
py_class(classname, ..., inherit = NULL)
}
\arguments{
\item{classname}{Character. Name of the class.}

\item{...}{Functions, attributes, or other definitions for the class.}

\item{inherit}{List. A list of Python class objects representing the
parent classes.}
}
\description{
This function works like \code{\link[reticulate:PyClass]{reticulate::PyClass()}}, but the \code{defs} argument
can now be provided as individual arguments (\code{...}) instead of a list.
}
\examples{

\dontrun{
Employee <- py_class("Employee", 
                     `__init__` = function(self, name, id) {
                       self$name <- name
                       self$id <- id
                       return(invisible(NULL))
                     },
                     get_email = function(self) {
                       paste0(self$name, "_", self$id, "@company.com")
                     })
Mike <- Employee("Mike", "1234")
Mike$get_email()
}


}
